requirements.txt
Django==5.1
djangorestframework>=3.15
django-filter>=24.2
Pillow>=10.0
psycopg2-binary>=2.9
